requirements.txt
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.10
arxiv==2.1.0
biopython==1.83
chromadb==0.4.22
pypdf==3.17.4
beautifulsoup4==4.12.3
requests==2.31.0
python-dotenv==1.0.0
streamlit==1.29.0
pydantic==2.5.3
numpy==1.24.4
pandas==2.1.4
matplotlib==3.8.2
openai>=1.10.0,<2.0.0